# openapi.yaml
openapi: 3.1.0
info:
  title: Meeting Management API
  version: 1.0.0
  description: MVP API for meeting scheduling, rooms, participants, attachments, email invitations, RSVP, and reminders.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Sign in with an existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
      responses:
        '200':
          description: Authenticated; access token expires in 30 minutes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [access_token, token_type, expires_in, user]
                        properties:
                          access_token: { type: string }
                          token_type: { type: string, const: Bearer }
                          expires_in: { type: integer, const: 1800 }
                          user: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /rooms:
    get:
      tags: [Rooms]
      summary: List available rooms for a time range
      parameters:
        - in: query
          name: from
          required: true
          schema: { type: string, format: date-time }
        - in: query
          name: to
          required: true
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Rooms not in maintenance or booked during the range
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Room' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /meetings:
    get:
      tags: [Meetings]
      summary: List meetings visible to the authenticated user
      description: from and to are required, the range may not exceed 31 days. Admin sees all meetings; other users see meetings they organize or attend. Client-supplied user IDs do not widen visibility.
      parameters:
        - in: query
          name: from
          required: true
          schema: { type: string, format: date-time }
        - in: query
          name: to
          required: true
          schema: { type: string, format: date-time }
        - in: query
          name: room_id
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Meetings in range
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Meeting' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Meetings]
      summary: Create a meeting
      description: Multipart request with JSON meeting metadata and optional files. Organizer comes from the access token and is automatically added as Accepted; at least one other internal active user is required. By default conflicts return 409 with up to three suggested times. After explicit user confirmation, retry with allow_conflicts=true to create despite room and/or participant conflicts; the created meeting includes warnings. Invitation email includes an ICS and participant-specific RSVP link for invitees (organizer has no RSVP token). A 15-minute email reminder is scheduled for each meeting participant.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/MeetingCreateMultipart' }
            encoding:
              meeting: { contentType: application/json }
              attachments: { contentType: application/pdf, application/vnd.openxmlformats-officedocument.wordprocessingml.document, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, application/vnd.openxmlformats-officedocument.presentationml.presentation }
      responses:
        '201':
          description: Meeting created; warnings may report invitation delivery failure or an explicitly overridden schedule conflict
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Meeting' }
                      meta:
                        allOf:
                          - $ref: '#/components/schemas/Meta'
                          - type: object
                            properties:
                              warnings:
                                type: array
                                items: { $ref: '#/components/schemas/Warning' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: Attachment exceeds 10 MB or request exceeds upload limit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '415':
          description: Attachment type is not PDF, DOCX, XLSX, or PPTX
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Schedule conflict; user may choose a suggested time or explicitly confirm creating despite conflicts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /meetings/{meetingId}:
    parameters:
      - $ref: '#/components/parameters/MeetingId'
    get:
      tags: [Meetings]
      summary: Get meeting details
      responses:
        '200':
          description: Meeting details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Meeting' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Meetings]
      summary: Update a meeting
      description: Organizer or Admin only. Multipart request with meeting metadata and optional attachments. Rechecks conflicts excluding this meeting; conflicts require explicit allow_conflicts confirmation. Editing time reschedules reminders. Notifications are sent after commit.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: '#/components/schemas/MeetingCreateMultipart' }
            encoding:
              meeting: { contentType: application/json }
              attachments: { contentType: application/pdf, application/vnd.openxmlformats-officedocument.wordprocessingml.document, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, application/vnd.openxmlformats-officedocument.presentationml.presentation }
      responses:
        '200':
          description: Meeting updated; warnings may report email delivery failure
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Meeting' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Meetings]
      summary: Cancel a meeting (soft cancel)
      responses:
        '200':
          description: Meeting cancelled; email delivery warning may be included
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          meeting_id: { type: integer }
                          status: { type: string, const: Cancelled }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    MeetingId:
      in: path
      name: meetingId
      required: true
      schema: { type: integer, minimum: 1 }
  schemas:
    Meta:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        requestId: { type: string }
    Success:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, const: true }
        data: {}
        meta: { $ref: '#/components/schemas/Meta' }
    Error:
      type: object
      required: [success, error, meta]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              description: For validation errors each item includes field, code, and localized message; conflict items identify the conflicting room/user/meeting.
              items:
                type: object
                properties:
                  field: { type: string }
                  code: { type: string }
                  message: { type: string }
                  entity_id: { type: integer }
                  entity_name: { type: string }
                  meeting_id: { type: integer }
            suggested_times:
              type: array
              maxItems: 3
              items:
                type: object
                required: [start_time, end_time]
                properties:
                  start_time: { type: string, format: date-time }
                  end_time: { type: string, format: date-time }
            override_allowed: { type: boolean }
        meta: { $ref: '#/components/schemas/Meta' }
    Warning:
      type: object
      properties:
        code: { type: string, enum: [INVITATION_EMAIL_FAILED, SCHEDULE_CONFLICT_OVERRIDDEN] }
        message: { type: string }
    User:
      type: object
      required: [user_id, full_name, email, role]
      properties:
        user_id: { type: integer }
        full_name: { type: string }
        email: { type: string, format: email }
        phone: { type: [string, 'null'] }
        role: { type: string, enum: [Admin, Organizer, Participant] }
    Room:
      type: object
      required: [room_id, room_name, capacity, status]
      properties:
        room_id: { type: integer }
        room_name: { type: string }
        capacity: { type: integer }
        location: { type: [string, 'null'] }
        status: { type: string, enum: [Available, Maintenance] }
    ParticipantInput:
      oneOf:
        - type: object
          required: [user_id]
          properties:
            user_id: { type: integer, minimum: 1 }
        - type: object
          required: [email]
          properties:
            email: { type: string, format: email }
    MeetingInput:
      type: object
      required: [title, start_time, end_time, participants]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: [string, 'null'] }
        start_time: { type: string, format: date-time, description: ISO 8601 timestamp; stored as UTC }
        end_time: { type: string, format: date-time, description: Must be after start_time and not in the past }
        room_id: { type: [integer, 'null'], minimum: 1 }
        meeting_link: { type: [string, 'null'], format: uri }
        participants:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: '#/components/schemas/ParticipantInput' }
        allow_conflicts: { type: boolean, default: false, description: Set true only after the user explicitly confirms creating despite the conflict warning. }
      description: At least one invitee other than the organizer. Room and meeting_link are individually optional per AC1; either or both may be supplied. Start must not be in the past; timestamps include an offset and are stored in UTC. Reminder is fixed to 15 minutes in V1.
    MeetingCreateMultipart:
      type: object
      required: [meeting]
      properties:
        meeting:
          $ref: '#/components/schemas/MeetingInput'
        attachments:
          type: array
          maxItems: 5
          items:
            type: string
            format: binary
            description: PDF, DOCX, XLSX, or PPTX; max 10 MB each.
      description: Multipart form with a JSON meeting part and up to five file parts. Organizer is excluded from the required invitee count and added automatically as Accepted.
    MeetingParticipant:
      type: object
      properties:
        user_id: { type: integer }
        full_name: { type: string }
        email: { type: string, format: email }
        status: { type: string, enum: [Pending, Accepted, Declined, Tentative] }
    Meeting:
      type: object
      properties:
        meeting_id: { type: integer }
        title: { type: string }
        description: { type: [string, 'null'] }
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time }
        room: { oneOf: [{ $ref: '#/components/schemas/Room' }, { type: 'null' }] }
        meeting_link: { type: [string, 'null'], format: uri }
        organizer: { $ref: '#/components/schemas/User' }
        participants: { type: array, items: { $ref: '#/components/schemas/MeetingParticipant' } }
        status: { type: string, enum: [Scheduled, In-progress, Completed, Cancelled] }
        created_at: { type: string, format: date-time }
        reminder_minutes_before: { type: integer, const: 15 }
        attachments:
          type: array
          items:
            type: object
            properties:
              attachment_id: { type: integer }
              file_name: { type: string }
              mime_type: { type: string }
              size_bytes: { type: integer }
    ErrorResponse:
      allOf: [{ $ref: '#/components/schemas/Error' }]
  responses:
    ValidationError:
      description: Invalid request fields or date range. Field messages include “Vui lòng nhập tên cuộc họp”, “Thời gian kết thúc phải sau thời gian bắt đầu”, and “Vui lòng chọn ít nhất 1 người tham gia”.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing, invalid, or expired access token / invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated user lacks permission or resource membership
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Meeting, room, or internal user does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Room/participant schedule conflict or invalid meeting state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }